# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Multi-Tenant Event Processing Platform
  description: "\nTenant-scoped event ingestion, querying, aggregation, search, and cached live summaries.\n\
    \nConsistency semantics, stated once and honoured everywhere:\n\n- `202 Accepted` means validated\
    \ and queued. It does NOT mean stored, searchable, or\n  reflected in any summary.\n- An event becomes\
    \ retrievable via `/events` before it becomes findable via `/events/search`.\n- `/events/stats/realtime`\
    \ may be stale by up to the configured freshness window.\n"
  version: 1.0.0
paths:
...
  /events/stats:
    get:
      tags:
      - stats
      summary: Counts grouped by type and bucket
      operationId: event_stats_events_stats_get
      parameters:
      - name: from
        in: query
        required: true
        schema:
          type: string
          format: date-time
          title: From
      - name: to
        in: query
        required: true
        schema:
          type: string
          format: date-time
          title: To
      - name: bucket
        in: query
        required: true
        schema:
          enum:
          - hour
          - day
          - week
          type: string
          title: Bucket
      - name: event_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Event Type
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /events/stats/realtime:
    get:
      tags:
      - stats
      summary: Lightweight live summary served from cache
      description: Always answers. Degrades to computation rather than failing if the cache is down.
      operationId: realtime_stats_events_stats_realtime_get
      parameters:
      - name: window
        in: query
        required: false
        schema:
          enum:
          - 5m
          - 1h
          - 24h
          type: string
          default: 1h
          title: Window
      - name: event_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Event Type
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RealtimeOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /events/search:
    get:
      tags:
      - search
      summary: Free-text search across allowlisted event metadata
      description: Terms are treated literally. Raw query DSL is not accepted.
      operationId: search_events_events_search_get
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 256
          title: Q
      - name: event_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Event Type
      - name: from
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: From
      - name: to
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: To
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    EventOut:
      properties:
        event_id:
          type: string
          title: Event Id
        tenant_id:
          type: string
          title: Tenant Id
        event_type:
          type: string
          title: Event Type
        occurred_at:
          type: string
          format: date-time
          title: Occurred At
        received_at:
          type: string
          format: date-time
          title: Received At
        schema_version:
          type: integer
          title: Schema Version
        anonymous_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Anonymous Id
        contact_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Contact Id
        session_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
        campaign_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Campaign Id
        channel:
          anyOf:
          - type: string
          - type: 'null'
          title: Channel
        source_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Url
        conversion:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Conversion
        consent:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Consent
        metadata:
          additionalProperties: true
          type: object
          title: Metadata
          default: {}
      type: object
      required:
      - event_id
      - tenant_id
      - event_type
      - occurred_at
      - received_at
      - schema_version
      title: EventOut
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    RangeAppliedOut:
      properties:
        start:
          type: string
          format: date-time
          title: Start
        end:
          type: string
          format: date-time
          title: End
        clamped:
          type: boolean
          title: Clamped
      type: object
      required:
      - start
      - end
      - clamped
      title: RangeAppliedOut
    RealtimeOut:
      properties:
        counts:
          additionalProperties:
            type: integer
          type: object
          title: Counts
        cached:
          type: boolean
          title: Cached
        age_seconds:
          type: integer
          title: Age Seconds
        degraded:
          type: boolean
          title: Degraded
      type: object
      required:
      - counts
      - cached
      - age_seconds
      - degraded
      title: RealtimeOut
    SearchOut:
      properties:
        items:
          items:
            $ref: '#/components/schemas/EventOut'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
        freshness_note:
          type: string
          title: Freshness Note
      type: object
      required:
      - items
      - total
      - freshness_note
      title: SearchOut
    StatsBucketOut:
      properties:
        bucket_start:
          type: string
          format: date-time
          title: Bucket Start
        event_type:
          type: string
          title: Event Type
        count:
          type: integer
          title: Count
      type: object
      required:
      - bucket_start
      - event_type
      - count
      title: StatsBucketOut
    StatsOut:
      properties:
        range_applied:
          $ref: '#/components/schemas/RangeAppliedOut'
        buckets:
          items:
            $ref: '#/components/schemas/StatsBucketOut'
          type: array
          title: Buckets
      type: object
      required:
      - range_applied
      - buckets
      title: StatsOut
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError